% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{Conduct an Analysis of Variance (ANOVA)}
\usage{
anova(df, IV, DV)
}
\arguments{
\item{df}{data frame}

\item{IV}{Independent/Grouping Variable}

\item{DV}{Dependent Variable}
}
\value{
ANOVA results
}
\description{
Conduct an Analysis of Variance (ANOVA)
}
