% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Convert variable to a Factor Data Type}
\usage{
convert_to_factor(df, var)
}
\arguments{
\item{df}{Data Frame}

\item{var}{Independent Variable (IV)}
}
\value{
Data Frame with IV as a factor data type
}
\description{
Test whether a variable is a Factor Data Type. If it isn't,
convert it to a Factor Data Type and return the data frame
back to the user. If it is, do nothing and return the data
frame as it is to the user.
}
