% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Compute the descriptive statistics}
\usage{
desc(df, IV, DV)
}
\arguments{
\item{df}{data frame}

\item{IV}{Independent/Grouping Variable}

\item{DV}{Dependent Variable}
}
\value{
Data Frame of Mean, Standard Deviation, and N by each IV/Grouping Variable
}
\description{
Compute the descriptive statistics
}
