% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{test_factor}
\alias{test_factor}
\title{Test whether the Independent Variable is a factor variable}
\usage{
test_factor(df, var)
}
\arguments{
\item{df}{Data Frame}

\item{var}{Independent Variable (IV)}
}
\value{
Data Frame with IV as a factor data type
}
\description{
Test whether the Independent Variable is a factor variable
}
