% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tidyCol}
\alias{tidyCol}
\title{Tidy up Column Names}
\usage{
tidyCol(col)
}
\arguments{
\item{col}{Column Name}
}
\value{
A tidied column name
}
\description{
Tidy up Column Names by removing underscores (_) and
periods (.) and replacing it with spaces ( ). Tidied column
names could then be used in visualizations
}
